+++ services/portfolio.py
@@ -1,0 +1,83 @@
+def fetch_portfolios():
+    """
+    Fetch the list of portfolios/stock symbols to monitor.
+    
+    Returns:
+        list: List of dictionaries containing stock symbols and portfolio info
+    """
+    return [
+                [
+            {
+                'symbol': 'VGT',
+                'portfolio': 'Cash'
+            },
+            {
+                'symbol': 'VOOG',
+                'portfolio': 'Cash'
+            },
+            {
+                'symbol': 'PLTR',
+                'portfolio': 'Cash'
+            },
+            {
+                'symbol': 'VDE',
+                'portfolio': 'Cash'
+            },
+            {
+                'symbol': 'NVDA',
+                'portfolio': 'Cash'
+            },
+            {
+                'symbol': 'META',
+                'portfolio': 'Cash'
+            },
+            {
+                'symbol': 'RY',
+                'portfolio': 'Cash'
+            },
+            {
+                'symbol': 'CDNS',
+                'portfolio': 'Cash'
+            },
+            {
+                'symbol': 'AISP',
+                'portfolio': 'Cash'
+            },
+            {
+                'symbol': 'USE',
+                'portfolio': 'Cash'
+            },
+            {
+                'symbol': 'BFM',
+                'portfolio': 'Cash'
+            },
+            {
+                'symbol': 'VUS',
+                'portfolio': 'Cash'
+            },
+            {
+                'symbol': 'TD',
+                'portfolio': 'Cash'
+            },
+            {
+                'symbol': 'AMZN',
+                'portfolio': 'Cash'
+            },
+            {
+                'symbol': 'SCHD',
+                'portfolio': 'Cash'
+            },
+            {
+                'symbol': 'TSLA',
+                'portfolio': 'Cash'
+            },
+            {
+                'symbol': 'GOOG',
+                'portfolio': 'Cash'
+            },
+            {
+                'symbol': 'BTC',
+                'portfolio': 'Cash'
+            }
+        ]
+    ]
